fix some linter warnings

+++ back/transcendence_django/back_game/game_geometry/vector.py
@@ -6,8 +6,8 @@ from back_game.game_geometry.position import Position
 class Vector(Position):
 
     def __init__(self, x: float, y: float):
-        self.x: int = x
-        self.y: int = y
+        self.x: float = x
+        self.y: float = y
 
     def magnitude(self) -> float:
         return math.sqrt(self.x**2 + self.y**2)
